refactor(xml): move to utils
